Fix the circular zod object import deps
