+++ web/packages/new/photos/services/user-details.ts
@@ -1,18 +1,18 @@
 import { authenticatedRequestHeaders, ensureOk } from "@/base/http";
+import { getKV, setKV } from "@/base/kv";
 import { apiURL, familyAppOrigin, paymentsAppOrigin } from "@/base/origins";
 import { ensure } from "@/utils/ensure";
-import { nullToUndefined } from "@/utils/transform";
+import { nullishToZero, nullToUndefined } from "@/utils/transform";
+import { getData, LS_KEYS, setLSUser } from "@ente/shared/storage/localStorage";
 import isElectron from "is-electron";
 import { z } from "zod";
-import type { UserDetails } from "./user";
-import { syncUserDetails, userDetailsSnapshot } from "./user";
 
 /**
  * Validity of the plan.
  */
 export type PlanPeriod = "month" | "year";
 
-export const Subscription = z.object({
+const Subscription = z.object({
     /**
      * Store-specific ID of the product ("plan") that the user has subscribed
      * to. e.g. if the user has subscribed to a plan using Stripe, then this
@@ -73,7 +73,7 @@ type FamilyMember = z.infer<typeof FamilyMember>;
  * Zod schema for details about the family plan (if any) that the user is a part
  * of.
  */
-export const FamilyData = z.object({
+const FamilyData = z.object({
     members: z.array(FamilyMember),
     /**
      * Family admin subscription storage capacity.
@@ -108,7 +108,7 @@ const Bonus = z.object({
  */
 export type Bonus = z.infer<typeof Bonus>;
 
-export const BonusData = z.object({
+const BonusData = z.object({
     /**
      * List of bonuses applied for the user.
      */
@@ -121,6 +121,147 @@ export const BonusData = z.object({
  * Information about bonuses applied for the user.
  */
 export type BonusData = z.infer<typeof BonusData>;
+
+/**
+ * Zod schema for {@link UserDetails}
+ */
+const UserDetails = z.object({
+    email: z.string(),
+    usage: z.number(),
+    fileCount: z.number().nullish().transform(nullishToZero),
+    subscription: Subscription,
+    familyData: FamilyData.nullish().transform(nullToUndefined),
+    storageBonus: z.number().nullish().transform(nullishToZero),
+    bonusData: BonusData.nullish().transform(nullToUndefined),
+});
+
+export type UserDetails = z.infer<typeof UserDetails>;
+
+/**
+ * Internal in-memory state shared by the functions in this module.
+ *
+ * This entire object will be reset on logout.
+ */
+class UserDetailsState {
+    /**
+     * Subscriptions to {@link UserDetails} updates attached using
+     * {@link userDetailsSubscribe}.
+     */
+    userDetailsListeners: (() => void)[] = [];
+
+    /**
+     * Snapshot of the {@link UserDetails} returned by the
+     * {@link userDetailsSnapshot} function.
+     */
+    userDetailsSnapshot: UserDetails | undefined;
+}
+
+/**
+ * State shared by the functions in this module. See {@link UserDetailsState}.
+ */
+let _state = new UserDetailsState();
+
+export const logoutUserDetails = () => {
+    _state = new UserDetailsState();
+};
+
+/**
+ * Read in the locally persisted settings into memory, otherwise initate a
+ * network requests to fetch the latest values (but don't wait for it to
+ * complete).
+ *
+ * This assumes that the user is already logged in.
+ */
+export const initUserDetailsOrTriggerSync = async () => {
+    const saved = await getKV("userDetails");
+    if (saved) {
+        setUserDetailsSnapshot(UserDetails.parse(saved));
+    } else {
+        void syncUserDetails();
+    }
+};
+
+/**
+ * A function that can be used to subscribe to updates to {@link UserDetails}.
+ *
+ * [Note: Snapshots and useSyncExternalStore]
+ *
+ * This subscribe function, along with {@link userDetailsSnapshot}, is meant to
+ * be used as arguments to React's {@link useSyncExternalStore}.
+ *
+ * @param callback A function that will be invoked whenever the result of
+ * {@link userDetailsSnapshot} changes.
+ *
+ * @returns A function that can be used to clear the subscription.
+ */
+export const userDetailsSubscribe = (onChange: () => void): (() => void) => {
+    _state.userDetailsListeners.push(onChange);
+    return () => {
+        _state.userDetailsListeners = _state.userDetailsListeners.filter(
+            (l) => l != onChange,
+        );
+    };
+};
+
+/**
+ * Return the last known, cached {@link UserDetails}.
+ *
+ * This, along with {@link userDetailsSubscribe}, is meant to be used as
+ * arguments to React's {@link useSyncExternalStore}.
+ */
+export const userDetailsSnapshot = () => _state.userDetailsSnapshot;
+
+const setUserDetailsSnapshot = (snapshot: UserDetails) => {
+    _state.userDetailsSnapshot = snapshot;
+    _state.userDetailsListeners.forEach((l) => l());
+};
+
+/**
+ * Fetch the user's details from remote and save them in local storage for
+ * subsequent lookup, and also update our in-memory snapshots.
+ */
+export const syncUserDetails = async () => {
+    const userDetails = await getUserDetailsV2();
+    await setKV("userDetails", userDetails);
+    setUserDetailsSnapshot(userDetails);
+
+    // TODO: The existing code used to also set the email for the local storage
+    // user whenever it updated the user details. I don't see why this would be
+    // needed though.
+    //
+    // Retaining the existing behaviour for now, except we throw. The intent is
+    // to remove this entire copy-over after a bit.
+    //
+    // Added Nov 2024, and can be removed after a while (tag: Migration).
+
+    const oldLSUser = getData(LS_KEYS.USER) as unknown;
+    const hasMatchingEmail =
+        oldLSUser &&
+        typeof oldLSUser == "object" &&
+        "email" in oldLSUser &&
+        typeof oldLSUser.email == "string" &&
+        oldLSUser.email == userDetails.email;
+
+    if (!hasMatchingEmail) {
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
+        await setLSUser({
+            ...getData(LS_KEYS.USER),
+            email: userDetails.email,
+        });
+        throw new Error("Email in local storage did not match user details");
+    }
+};
+
+/**
+ * Fetch user details from remote.
+ */
+export const getUserDetailsV2 = async () => {
+    const res = await fetch(await apiURL("/users/details/v2"), {
+        headers: await authenticatedRequestHeaders(),
+    });
+    ensureOk(res);
+    return UserDetails.parse(await res.json());
+};
 
 /**
  * Zod schema for an individual plan received in the list of plans.
